#191950 project settings interface renew
