+++ src/components/UiButton.vue
@@ -155,7 +155,8 @@ $primary-button-size: 14px;
   padding: 8px 24px;
   position: relative;
   text-align: center;
-  transition: color 0.2s ease-out, background-color 0.2s ease-out, box-shadow 0.2s ease-out;
+  transition: color 0.2s ease-out, background-color 0.2s ease-out,
+    box-shadow 0.2s ease-out;
 
   &:not(._disabled):not(._transparent):before {
     bottom: 1px;
@@ -239,6 +240,14 @@ $primary-button-size: 14px;
   }
   &._gray {
     @include base-button($gray-button-color);
+
+    &._transparent {
+      color: #000000;
+
+      &:hover {
+        color: $gray-button-color;
+      }
+    }
   }
   &._light-gray {
     @include base-button(
