+++ src/pages/ProjectSettingsPage.vue
@@ -1,29 +1,72 @@
 <script>
-import { mapState, mapMutations } from 'vuex';
-import ProjectFormSettings from '@/components/ProjectFormSettings.vue';
+import { forEach, get, set } from 'lodash-es';
+import { mapState, mapMutations, mapActions } from 'vuex';
+import { required } from 'vuelidate/lib/validators';
 
 export default {
   name: 'ProjectSettingsPage',
-  components: {
-    ProjectFormSettings,
-  },
-  props: {
-    uploadImage: {
-      type: Function,
-      required: true,
-    },
+
+  data() {
+    return {
+      cover: {
+        images: { en: '', ru: '' },
+        use_one_for_all: true,
+      },
+      langs: ['en', 'ru'],
+      fullDescription: {
+        en: '',
+        ru: '',
+      },
+      shortDescription: {
+        en: '',
+        ru: '',
+      },
+    };
   },
 
   computed: {
     ...mapState('Project', ['project', 'currencies']),
   },
 
+  validations() {
+    const rules = {
+      project: {
+        name: {
+          en: {
+            required,
+          },
+        },
+      },
+    };
+
+    return rules;
+  },
+
   methods: {
+    ...mapActions(['uploadImage']),
     ...mapMutations('Project', { setCurrencies: 'currencies' }),
 
+    updateLangs(langs) {
+      this.langs = langs;
+
+      const langFields = [
+        'project.name',
+        'fullDescription',
+        'shortDescription',
+      ];
+
+      langFields.forEach((keypath) => {
+        const newFieldValue = {};
+        forEach(langs, (lang) => {
+          newFieldValue[lang] = get(this, `${keypath}.${lang}`) || '';
+        });
+        set(this, keypath, newFieldValue);
+      });
+    },
+
     handleSave() {
-      const isValid = this.$refs.form.chekIsFormValid();
-      if (isValid) {
+      this.$v.$touch();
+      if (!this.$v.$invalid) {
         this.$emit('save');
       }
     },
@@ -38,20 +81,85 @@ export default {
     <span slot="description">
       Setup main scope of project parameters here.
     </span>
-    <!-- <PictureClock slot="picture" /> -->
   </UiPageHeaderFrame>
 
   <UiPanel>
-    <ProjectFormSettings
-      ref="form"
-      v-bind="{
-        uploadImage,
-        currencies,
-        project
-      }"
-      v-on="$listeners"
-      @currenciesChange="setCurrencies"
+    <UiHeader
+      :hasMargin="true"
+      level="3"
+    >
+      Localizations
+    </UiHeader>
+    <UiText indentBottom="small">
+      Choose complete set of supported languages for all products and descriptions.
+    </UiText>
+    <UiLangsMainHub
+      :langs="langs"
+      @change="updateLangs"
     />
+  </UiPanel>
+
+  <UiPanel>
+    <UiSwitchBox
+      class="localization-switch"
+      v-model="cover.use_one_for_all"
+    >
+      Use single cover for all localizations
+    </UiSwitchBox>
+
+    <UiLangsImageUpload
+      class="section"
+      :uploadImage="uploadImage"
+      :isLocalizationEnabled="!cover.use_one_for_all"
+      v-model="cover.images"
+    />
+
+    <div class="section">
+      <UiHeader
+        :hasMargin="true"
+        level="3"
+      >
+        Project description
+      </UiHeader>
+      <UiText>
+        Fill-in the unique name and localised project descriptions here.
+        It will be mentioned in payment form and different customer correspondence.
+      </UiText>
+      <UiLangTextField
+        :value="project.name"
+        :langs="langs"
+        label="Project name"
+        v-bind="$getValidatedFieldProps('project.name.en')"
+      />
+      <UiLangTextField
+        :value="fullDescription"
+        :langs="langs"
+        label="Full description"
+      />
+      <UiLangTextField
+        :value="shortDescription"
+        :langs="langs"
+        label="Short description"
+      />
+    </div>
+
+    <div class="section">
+      <UiHeader
+        :hasMargin="true"
+        level="3"
+      >
+        Currencies
+      </UiHeader>
+      <UiText indentBottom="small">
+        Choose a fixed set of currencies which will be actual for all products in this project.
+        <span class="bold">USD is default currency</span>.
+      </UiText>
+
+      <UiCurrenciesMainHub
+        :currencies="currencies"
+        @change="$emit('currenciesChange', setCurrencies)"
+      />
+    </div>
 
     <div class="controls">
       <UiButton
@@ -66,6 +174,25 @@ export default {
 
 
 <style lang="scss" scoped>
+.section {
+  margin-bottom: 32px;
+}
+
+.localization-switch {
+  margin-bottom: 20px;
+}
+
+.switch-box {
+  margin-left: 16px;
+}
+
+.bold {
+  font-weight: 500;
+}
+
+.generate-button {
+  margin-left: 30px;
+}
 .controls {
   display: flex;
   justify-content: flex-end;
